Language Code Update, main branch (2026.06.04.) (#1342)
* Use vecmem-check-language consistently throughout the project.

* Make Alpaka make use of traccc-compiler-options-hip.

* Fixing warnings coming from Alpaka+HIP.

Unfortunately so far I didn't manage to reproduce these locally,
so this might not fix all the issues yet...

---------

+++ core/include/traccc/fitting/kalman_filter/kalman_actor.hpp
@@ -475,14 +475,16 @@ struct kalman_actor : detray::base_actor {
             // Abort if the Kalman update fails
             if (actor_state.fit_result != kalman_fitter_status::SUCCESS) {
                 if (actor_state.backward_mode) {
-                    TRACCC_ERROR_DEVICE("Abort backward fit: KF status %d",
-                                        actor_state.fit_result);
+                    TRACCC_ERROR_DEVICE(
+                        "Abort backward fit: KF status %d",
+                        static_cast<int>(actor_state.fit_result));
                     TRACCC_ERROR_HOST(
                         "Abort backward fit: "
                         << fitter_debug_msg{actor_state.fit_result}());
                 } else {
-                    TRACCC_ERROR_DEVICE("Abort forward fit: KF status %d",
-                                        actor_state.fit_result);
+                    TRACCC_ERROR_DEVICE(
+                        "Abort forward fit: KF status %d",
+                        static_cast<int>(actor_state.fit_result));
                     TRACCC_ERROR_HOST("Abort forward fit: " << fitter_debug_msg{
                                           actor_state.fit_result}());
                 }
